Add Azure Communication Services Emailer
Health check is not implemented, because the SDK has no sandbox option.

I tried to create a workaround with the cancellation token, but then we only receive a status cancelled, which is not helpful.

When sending a message, it is possible to either wait until the operation was started, or also completed. We could create a message in the health check, wait until it started (without an exception), and then cancel the operation.

Depending on the service communication delays, this might send a real mail, which might not be what we want.

If we don't care, we can create a mailbox, which just discards our health check emails. However, this option will eat into our daily send limits.